+++ CHANGES.rst
@@ -14,7 +14,9 @@ New features:
 
 Bug fixes:
 
-- *add item here*
+- Fix bug #56 where ``srcset`` generation failed on no given width or height if there was no sclae given.
+  https://github.com/plone/plone.namedfile/pull/56
+  [jensens]
 
 
 4.2.2 (2017-07-03)